Use Enum for K8s environment
