+++ infra/fridge/enums/__init__.py
@@ -1,0 +1,25 @@
+from enum import Enum, unique
+
+
+@unique
+class K8sEnvironment(Enum):
+    AKS = "AKS"
+    DAWN = "Dawn"
+
+
+@unique
+class PodSecurityStandard(Enum):
+    RESTRICTED = {"pod-security.kubernetes.io/enforce": "restricted"}
+    PRIVILEGED = {"pod-security.kubernetes.io/enforce": "privileged"}
+
+
+@unique
+class TlsEnvironment(Enum):
+    STAGING = "staging"
+    PRODUCTION = "production"
+
+
+tls_issuer_names = {
+    TlsEnvironment.STAGING: "letsencrypt-staging",
+    TlsEnvironment.PRODUCTION: "letsencrypt-prod",
+}
